+++ cmd/tidio/router_test.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"net/http"
 	"testing"
 
 	"github.com/gregoryv/asserter"
@@ -8,7 +9,11 @@ import (
 
 func Test_router(t *testing.T) {
 	assert := asserter.New(t)
-	apikeys := make(map[string]string)
+	apikeys := map[string]string{
+		"KEY": "john",
+	}
+	headers := http.Header{}
+
 	exp := assert().ResponseFrom(NewRouter(apikeys))
 	exp.StatusCode(200, "GET", "/api", nil)
 	exp.Contains("revision", "GET", "/api")
@@ -16,5 +21,13 @@ func Test_router(t *testing.T) {
 	exp.Contains("resources", "GET", "/api")
 
 	exp.StatusCode(401, "GET", "/api/timesheets/")
+	headers = http.Header{}
+	headers.Set("Authorization", "NO SUCH KEY")
+	exp.StatusCode(401, "GET", "/api/timesheets/", headers)
 	exp.StatusCode(401, "POST", "/api/timesheets/")
+
+	// authenticated
+	headers = http.Header{}
+	headers.Set("Authorization", "KEY")
+	exp.StatusCode(200, "GET", "/api/timesheets/", headers)
 }